Refactor type definitions so that they can be properly picked up
