+++ test/helpers/create-test-server.d.ts
@@ -1,15 +1,5 @@
 declare module 'create-test-server' {
-	import {Express} from 'express';
-
-	export interface TestServer extends Express {
-		caCert: any;
-		port: number;
-		url: string;
-		sslPort: number;
-		sslUrl: string;
-
-		close(): Promise<void>;
-	}
+	import {TestServer} from './types';
 
 	function createTestServer(options: any): Promise<TestServer>;
